Beta 02
Fixed clients that join a game that has one or more disconnected
players, from not properly being given the info on those disconnected
players boxes. (This mean if the game was over, the client would think
it was actually still running.)

Effectively I stopped removing clients from the hashmap or whatever. I'm
curious how this will affect memory in the future, we may need to just
update the protocol to allow for a single "system" user, but that seemed
like a lot of work.

Also the debug magenta is no longer used for disconnected players, black
is. That magenta really is meant for catching bugs...